@@ -1,62 +1,92 @@
 {
   "name": "ngx-ui-scroll-app",
   "version": "0.0.0",
-  "license": "MIT",
+  "description": "Angular Demo App for ngx-ui-scroll",
+  "main": "./bundles/ngx-ui-scroll.umd.js",
+  "module": "./fesm5/ngx-ui-scroll.js",
+  "es2015": "./fesm2015/ngx-ui-scroll.js",
+  "esm5": "./esm5/ngx-ui-scroll.js",
+  "esm2015": "./esm2015/ngx-ui-scroll.js",
+  "fesm5": "./fesm5/ngx-ui-scroll.js",
+  "fesm2015": "./fesm2015/ngx-ui-scroll.js",
+  "typings": "./ngx-ui-scroll.d.ts",
   "scripts": {
-    "start": "ng serve --delete-output-path=false",
-    "build": "node build.js",
+    "start": "npm-run-all --parallel start-api-server start-dev-server",
+    "start-dev-server": "ng serve --proxy-config proxy.conf.json",
+    "start-api-server": "node ./server/index.js",
     "test": "karma start",
-    "install-package": "npm run build && npm pack ./dist && npm install ngx-ui-scroll-0.0.0.tgz --no-save",
-    "start-package": "npm run install-package && npm start"
+    "build-app": "ng build --prod --aot --source-map",
+    "preinstall": "cd server && npm install",
+    "postinstall": "npm run build-app",
+    "pack:install": "npm run build && npm pack ./dist && npm install ngx-ui-scroll-1.0.1.tgz --no-save",
+    "pack:start": "npm run pack:install && npm start",
+    "build": "node build.js",
+    "publish:lib": "npm run build && npm publish ./dist",
+    "publish:lib:next": "npm run build && npm publish --tag next ./dist"
   },
-  "private": true,
+  "license": "MIT",
   "dependencies": {
-    "@angular/animations": "^5.2.0",
-    "@angular/common": "^5.2.0",
-    "@angular/core": "^5.2.0",
-    "@angular/forms": "^5.2.0",
-    "@angular/http": "^5.2.0",
-    "@angular/platform-browser": "^5.2.0",
-    "@angular/platform-browser-dynamic": "^5.2.0",
-    "@angular/router": "^5.2.0",
-    "core-js": "^2.5.3",
-    "rxjs": "^5.5.6",
-    "zone.js": "^0.8.20"
+    "@angular/animations": "^7.0.3",
+    "@angular/common": "^7.0.3",
+    "@angular/core": "^7.0.3",
+    "@angular/forms": "^7.0.3",
+    "@angular/http": "^7.0.3",
+    "@angular/platform-browser": "^7.0.3",
+    "@angular/platform-browser-dynamic": "^7.0.3",
+    "@angular/router": "^7.0.3",
+    "core-js": "^2.5.7",
+    "ngx-bootstrap": "^3.0.1",
+    "rxjs": "^6.3.3",
+    "tslib": "^1.9.0",
+    "zone.js": "^0.8.26"
+  },
+  "peerDependencies": {
+    "rxjs": ">= 6.0.0",
+    "@angular/common": ">= 7.0.0",
+    "@angular/core": ">= 7.0.0"
   },
   "devDependencies": {
-    "@angular/cli": "^1.6.4",
-    "@angular/compiler": "^5.2.0",
-    "@angular/compiler-cli": "^5.2.0",
-    "@angular/language-service": "^5.2.0",
-    "@types/jasmine": "^2.5.54",
-    "@types/jasminewd2": "~2.0.2",
-    "@types/node": "^6.0.95",
-    "chalk": "^2.3.0",
-    "codelyzer": "~3.2.0",
-    "cpx": "^1.5.0",
-    "jasmine-core": "~2.6.2",
-    "karma": "~1.7.0",
-    "karma-chrome-launcher": "~2.1.1",
+    "@angular-devkit/build-angular": "^0.8.1",
+    "@angular/cli": "^7.0.5",
+    "@angular/compiler": "^7.0.3",
+    "@angular/compiler-cli": "^7.0.3",
+    "@types/jasmine": "2.8.9",
+    "@types/jasminewd2": "~2.0.3",
+    "@types/node": "10.12.0",
+    "chalk": "2.4.1",
+    "codelyzer": "4.5.0",
+    "istanbul-instrumenter-loader": "3.0.1",
+    "jasmine-core": "3.2.1",
+    "karma": "3.0.0",
+    "karma-chrome-launcher": "2.2.0",
     "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^1.3.1",
+    "karma-coverage-istanbul-reporter": "2.0.4",
     "karma-firefox-launcher": "^1.1.0",
-    "karma-ie-launcher": "^1.0.0",
-    "karma-jasmine": "^1.1.1",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "karma-sourcemap-loader": "^0.3.7",
+    "karma-jasmine": "1.1.2",
+    "karma-sourcemap-loader": "0.3.7",
     "karma-spec-reporter": "0.0.32",
-    "karma-webpack": "^2.0.9",
-    "renamer": "^0.6.1",
-    "rollup": "^0.53.4",
-    "rollup-plugin-license": "^0.5.0",
-    "rollup-plugin-node-resolve": "^3.0.2",
-    "rollup-plugin-sourcemaps": "^0.4.2",
-    "shelljs": "^0.7.8",
-    "ts-loader": "^3.2.0",
-    "ts-node": "~3.2.0",
-    "tslint": "^5.9.1",
-    "typescript": "^2.5.3",
-    "uglify-js": "^3.3.7",
-    "webpack": "^3.10.0"
+    "karma-webpack": "4.0.0-rc.2",
+    "node-sass": "4.9.4",
+    "npm-run-all": "^4.1.3",
+    "reflect-metadata": "0.1.12",
+    "rollup": "0.66.6",
+    "rollup-plugin-license": "0.7.0",
+    "rollup-plugin-node-resolve": "3.4.0",
+    "rollup-plugin-sourcemaps": "0.4.2",
+    "shelljs": "0.8.2",
+    "source-map-loader": "0.2.4",
+    "ts-loader": "5.2.2",
+    "tslint": "5.11.0",
+    "tslint-angular": "1.1.2",
+    "typescript": "~3.1.6",
+    "uglify-js": "3.4.9",
+    "webpack": "4.21.0",
+    "angular2-template-loader": "0.6.2",
+    "raw-loader": "0.5.1",
+    "css-loader": "1.0.0",
+    "sass-loader": "7.1.0",
+    "to-string-loader": "1.1.5",
+    "tsickle": "0.33.0",
+    "webpack-filter-warnings-plugin": "1.2.0"
   }
 }
