@@ -1,101 +1,194 @@
-import { Observable } from 'rxjs/Observable';
+import { Subscription, BehaviorSubject } from 'rxjs';
 
-import { Datasource, Direction } from './interfaces/index';
-import { Settings } from './classes/settings';
-import { Viewport } from './classes/viewport';
-import { Buffer } from './classes/buffer';
-import { FetchModel } from './classes/fetch';
-import { ClipModel } from './classes/clip';
-
-import { checkDatasource } from './utils/index';
+import { UiScrollComponent } from '../ui-scroll.component';
+import { Scroller } from './scroller';
+import { Process, ProcessStatus as Status, ProcessSubject } from './interfaces/index';
+import { Init, Scroll, Reload, Start, PreFetch, Fetch, PostFetch, Render, Clip, Adjust, End } from './processes/index';
 
 export class Workflow {
 
-  private observer;
-  public resolver: Observable<any>;
+  scroller: Scroller;
+  process$: BehaviorSubject<ProcessSubject>;
+  cyclesDone: number;
 
-  public bindData: Function;
-  public datasource: Datasource;
-  public settings: Settings;
-  public viewport: Viewport;
-  public buffer: Buffer;
+  readonly context: UiScrollComponent;
+  readonly onScrollHandler: EventListener;
+  private itemsSubscription: Subscription;
+  private workflowSubscription: Subscription;
+  private scrollEventOptions: any;
 
-  public count = 0;
-  public countDone = 0;
-  public pending: boolean;
-  public direction: Direction;
-  public next: boolean;
-  public fetch: FetchModel;
-  public clip: ClipModel;
+  constructor(context: UiScrollComponent) {
+    this.context = context;
+    this.scroller = new Scroller(this.context, this.callWorkflow.bind(this));
+    this.process$ = new BehaviorSubject(<ProcessSubject>{
+      process: Process.init,
+      status: Status.start
+    });
+    this.cyclesDone = 0;
+    this.onScrollHandler = event => Scroll.run(this.scroller, { event });
 
-  constructor(context) {
-    this.resolver = Observable.create(_observer => this.observer = _observer);
-
-    this.bindData = () => {
-      this.next = true;
-      context.changeDetector.markForCheck();
-    };
-    this.datasource = checkDatasource(context.datasource);
-
-    this.settings = new Settings(context.datasource.settings);
-    this.viewport = new Viewport(context.elementRef, this.settings);
-    this.buffer = new Buffer();
-
-    this.fetch = new FetchModel();
-    this.clip = new ClipModel();
+    if (this.scroller.settings.initializeDelay) {
+      setTimeout(() => this.init(), this.scroller.settings.initializeDelay);
+    } else {
+      this.init();
+    }
   }
 
-  reset() {
-    this.pending = false;
-    this.direction = null;
-    this.next = false;
-    this.fetch.reset();
-    this.clip = new ClipModel();
+  init() {
+    this.scroller.init();
+    this.scroller.logger.stat('initialization');
+    this.initListeners();
   }
 
-  start(direction?: Direction) {
-    this.count++;
-    this.log(`---=== Workflow ${this.count} run`);
-    this.reset();
-    this.pending = true;
-    this.direction = direction || null;
-    return Promise.resolve(this);
+  initListeners() {
+    const scroller = this.scroller;
+    scroller.logger.log(() => `uiScroll Workflow listeners are being initialized`);
+    this.itemsSubscription = scroller.buffer.$items.subscribe(items => this.context.items = items);
+    this.workflowSubscription = this.process$.subscribe(this.process.bind(this));
+    this.initScrollEventListener();
   }
 
-  continue() {
-    return Promise.resolve(this);
+  initScrollEventListener() {
+    let passiveSupported = false;
+    try {
+      window.addEventListener(
+        'test', <EventListenerOrEventListenerObject>{}, Object.defineProperty({}, 'passive', {
+          get: () => passiveSupported = true
+        })
+      );
+    } catch (err) {
+    }
+    this.scrollEventOptions = passiveSupported ? { passive: false } : false;
+    this.scroller.viewport.scrollEventElement.addEventListener(
+      'scroll', this.onScrollHandler, this.scrollEventOptions
+    );
   }
 
-  finalize() { // stop 1 cycle
+  detachScrollEventListener() {
+    this.scroller.viewport.scrollEventElement.removeEventListener(
+      'scroll', this.onScrollHandler, this.scrollEventOptions
+    );
   }
 
-  end() {
-    this.pending = false;
-    this.countDone++;
-    this.viewport.saveScrollPosition();
-    this.finalize();
+  process(data: ProcessSubject) {
+    const scroller = this.scroller;
+    const { status, payload } = data;
+    this.scroller.logger.logProcess(data);
+    if (status === Status.error) {
+      End.run(scroller, payload);
+      return;
+    }
+    switch (data.process) {
+      case Process.init:
+        if (status === Status.start) {
+          Init.run(scroller);
+        }
+        if (status === Status.next) {
+          Start.run(scroller, payload);
+        }
+        break;
+      case Process.reload:
+        if (status === Status.start) {
+          Reload.run(scroller, payload);
+        }
+        if (status === Status.next) {
+          Init.run(scroller);
+        }
+        break;
+      case Process.scroll:
+        if (status === Status.next) {
+          if (!(payload && payload.keepScroll)) {
+            Init.run(scroller, payload);
+          } else {
+            Start.run(scroller, payload);
+          }
+        }
+        break;
+      case Process.start:
+        if (status === Status.next) {
+          PreFetch.run(scroller);
+        }
+        break;
+      case Process.preFetch:
+        if (status === Status.done) {
+          End.run(scroller);
+        }
+        if (status === Status.next) {
+          Fetch.run(scroller);
+        }
+        break;
+      case Process.fetch:
+        if (status === Status.next) {
+          PostFetch.run(scroller);
+        }
+        break;
+      case Process.postFetch:
+        if (status === Status.next) {
+          Render.run(scroller);
+        }
+        if (status === Status.done) {
+          End.run(scroller);
+        }
+        break;
+      case Process.render:
+        if (status === Status.next) {
+          if (scroller.settings.infinite) {
+            Adjust.run(scroller);
+          } else {
+            Clip.run(scroller);
+          }
+        }
+        break;
+      case Process.clip:
+        if (status === Status.next) {
+          Adjust.run(scroller);
+        }
+        break;
+      case Process.adjust:
+        if (status === Status.done) {
+          End.run(scroller);
+        }
+        break;
+      case Process.end:
+        if (status === Status.next) {
+          if (payload && payload.keepScroll) {
+            Scroll.run(scroller);
+          } else {
+            Start.run(scroller, payload);
+          }
+        }
+        if (status === Status.done) {
+          this.done();
+        }
+        break;
+    }
+  }
+
+  callWorkflow(processSubject: ProcessSubject) {
+    this.process$.next(processSubject);
   }
 
   done() {
-    this.log(`---=== Workflow ${this.count} done`);
-    this.end();
-    this.observer.next(this.next);
-  }
-
-  fail(error: any) {
-    this.log(`---=== Workflow ${this.count} fail`);
-    this.end();
-    this.observer.error(error);
+    const { state } = this.scroller;
+    this.cyclesDone++;
+    state.workflowCycleCount = this.cyclesDone + 1;
+    state.isInitialWorkflowCycle = false;
+    state.workflowPending = false;
+    if (state.scrollState.scrollTimer === null) {
+      state.isLoading = false;
+    }
+    this.finalize();
   }
 
   dispose() {
-    this.observer.complete();
+    this.detachScrollEventListener();
+    this.process$.complete();
+    this.workflowSubscription.unsubscribe();
+    this.itemsSubscription.unsubscribe();
+    this.scroller.dispose();
   }
 
-  log(...args) {
-    if (this.settings.debug) {
-      console.log.apply(this, args);
-    }
+  finalize() {
   }
 
 }
