@@ -1,34 +1,48 @@
+// Karma configuration for Unit testing
+
+const path = require('path');
+const FilterWarningsPlugin = require('webpack-filter-warnings-plugin');
+
 module.exports = function (config) {
-  config.set({
+
+  const configuration = {
+
+    // base path that will be used to resolve all patterns (eg. files, exclude)
     basePath: '',
+
+    // frameworks to use
+    // available frameworks: https://npmjs.org/browse/keyword/karma-adapter
     frameworks: ['jasmine'],
+
     plugins: [
       require('karma-jasmine'),
-      require('karma-webpack'),
-      require('karma-spec-reporter'),
-      require('karma-ie-launcher'),
       require('karma-chrome-launcher'),
       require('karma-firefox-launcher'),
+      require('karma-webpack'),
       require('karma-sourcemap-loader'),
-      require('karma-jasmine-html-reporter'),
-      require('karma-coverage-istanbul-reporter')
+      require('karma-spec-reporter'),
+      require('karma-coverage-istanbul-reporter'),
+      require('istanbul-instrumenter-loader')
     ],
+
+    // list of files / patterns to load in the browser
     files: [
       'tests/miscellaneous/styles.css',
-      { pattern: './tests/miscellaneous/entry-files.js', watched: false }
+      { pattern: './tests/_index.js', watched: false }
     ],
-    client: {
-      clearContext: false // leave Jasmine Spec Runner output visible in browser
-    },
-    coverageIstanbulReporter: {
-      reports: [ 'html', 'lcovonly' ],
-      fixWebpackSourcePaths: true
+
+    // list of files to exclude
+    exclude: [],
+
+    // preprocess matching files before serving them to the browser
+    // available preprocessors: https://npmjs.org/browse/keyword/karma-preprocessor
+    preprocessors: {
+      './tests/_index.js': ['webpack', 'sourcemap']
     },
 
-    preprocessors: {
-      './tests/miscellaneous/entry-files.js': ['webpack', 'sourcemap']
-    },
+    // webpack
     webpack: {
+      mode: 'development',
       resolve: {
         extensions: ['.ts', '.js']
       },
@@ -36,27 +50,83 @@ module.exports = function (config) {
         rules: [
           {
             test: /\.ts/,
-            loaders: ['ts-loader'],
+            use: [
+              { loader: 'ts-loader' },
+              { loader: 'angular2-template-loader' },
+              { loader: 'source-map-loader' }
+            ],
             exclude: /node_modules/
+          },
+          {
+            test: /\.html$/,
+            use: 'raw-loader'
+          },
+          {
+            test: /\.css$/,
+            use: [
+              { loader: 'to-string-loader' },
+              { loader: 'css-loader' }]
+          },
+          {
+            test: /\.scss$/,
+            use: [
+              { loader: 'raw-loader' },
+              { loader: 'sass-loader' }
+            ]
+          },
+          {
+            enforce: 'post',
+            test: /\.ts/,
+            use: [
+              {
+                loader: 'istanbul-instrumenter-loader',
+                options: { esModules: true }
+              }
+            ],
+            exclude: [
+              /\.spec.ts/,
+              /node_modules/
+            ]
           }
         ],
         exprContextCritical: false
       },
+      devtool: 'inline-source-map',
       performance: { hints: false },
-      devtool: 'inline-source-map'
+      /* Workaround for https://github.com/angular/angular/issues/21560 */
+      plugins: [
+        new FilterWarningsPlugin({
+          exclude: /System.import/
+        })
+      ]
     },
+
     webpackServer: {
       noInfo: true
     },
 
-    reporters: ['spec'],
+    // test results reporter to use
+    // possible values: 'dots', 'progress'
+    // available reporters: https://npmjs.org/browse/keyword/karma-reporter
+    reporters: ['spec', 'coverage-istanbul'],
+
+    coverageIstanbulReporter: {
+      reports: ['html', 'lcovonly'],
+      dir: path.join(__dirname, 'coverage'),
+      fixWebpackSourcePaths: true
+    },
+
     port: 9876,
     colors: true,
     logLevel: config.LOG_ERROR,
-    autoWatch: process.env.TRAVIS ? false : true,
+    autoWatch: !process.env.TRAVIS,
     browsers: process.env.TRAVIS ?
       ['Firefox'] :
       [process.platform === 'linux' ? 'Chromium' : 'Chrome'],
-    singleRun: process.env.TRAVIS ? true : false
-  });
+    singleRun: !!process.env.TRAVIS
+
+  };
+
+  config.set(configuration);
+
 };
