@@ -1,2 +1,3 @@
 export { UiScrollModule } from './src/ui-scroll.module';
-export { Datasource } from './src/component/interfaces/datasource';
+export { Datasource as IDatasource } from './src/component/interfaces/datasource';
+export { Datasource } from './src/component/classes/datasource';
