@@ -4,19 +4,20 @@ import sourcemaps from 'rollup-plugin-sourcemaps';
 const globals = {
   '@angular/core': 'ng.core',
   '@angular/common': 'ng.common',
-  'rxjs/Observable': 'Rx',
-  'rxjs/BehaviorSubject': 'Rx'
+  'rxjs': 'rxjs',
+  'rxjs/operators': 'rxjs.operators'
 };
 
 export default {
   external: Object.keys(globals),
   plugins: [resolve(), sourcemaps()],
-  onwarn: () => { return },
+  onwarn: () => null,
   output: {
     format: 'umd',
     name: 'ng.ngxUiScroll',
-    globals: globals,
+    globals,
     sourcemap: true,
-    exports: 'named'
+    exports: 'named',
+    amd: { id: 'ngx-ui-scroll' }
   }
-}
+};
