@@ -1,4 +1,6 @@
-export const checkDatasource = (datasource) => {
+import { Datasource as IDatasource } from '../interfaces/index';
+
+export const checkDatasource = (datasource: IDatasource) => {
   if (!datasource) {
     throw new Error('No datasource provided');
   }
@@ -8,7 +10,7 @@ export const checkDatasource = (datasource) => {
   if (typeof datasource.get !== 'function') {
     throw new Error('Datasource get is not a function');
   }
-  if (datasource.get.length < 2) {
+  if ((<Function>(datasource.get)).length < 2) {
     throw new Error('Datasource get method invalid signature');
   }
   return datasource;
