@@ -1,56 +1,56 @@
 import {
   Component, OnInit, OnDestroy,
-  TemplateRef, ElementRef, Renderer2,
+  TemplateRef, ElementRef,
   ChangeDetectionStrategy, ChangeDetectorRef
 } from '@angular/core';
 
-import { WorkflowRunner } from './component/runner';
-import { Datasource } from './component/interfaces/index';
+import { Workflow } from './component/workflow';
+import { Datasource as IDatasource } from './component/interfaces/index';
+import { Datasource } from './component/classes/datasource';
 import { Item } from './component/classes/item';
 
 @Component({
-  selector: 'ui-scroll',
+  selector: '[ui-scroll]',
   changeDetection: ChangeDetectionStrategy.OnPush,
-  template: `
-<div data-padding-backward></div>
-<div *ngFor="let item of items" id="{{item.nodeId}}">
-  <div [style.position]="item.invisible ? 'fixed' : null" [style.left]="item.invisible ? '-99999px' : null" >
-    <ng-template
-      [ngTemplateOutlet]="template"
-      [ngTemplateOutletContext]="{
-        $implicit: item.data,
-        index: item.$index
-     }">
-    </ng-template>
-  </div>
-</div>
-<div data-padding-forward></div>
-`
+  template: `<div data-padding-backward></div><div
+  *ngFor="let item of items"
+  [attr.data-sid]="item.nodeId"
+  [style.position]="item.invisible ? 'fixed' : null"
+  [style.left]="item.invisible ? '-99999px' : null"
+><ng-template
+  [ngTemplateOutlet]="template"
+  [ngTemplateOutletContext]="{
+    $implicit: item.data,
+    index: item.$index,
+    odd: item.$index % 2,
+    even: !(item.$index % 2)
+ }"
+></ng-template></div><div data-padding-forward></div>`
 })
 export class UiScrollComponent implements OnInit, OnDestroy {
 
   // come from the directive
+  public version: string;
   public template: TemplateRef<any>;
-  public datasource: Datasource;
+  public datasource: IDatasource | Datasource;
 
   // use in the template
   public items: Array<Item>;
 
   // Component-Workflow integration
-  public workflowRunner: WorkflowRunner;
+  public workflow: Workflow;
 
   constructor(
-    private changeDetector: ChangeDetectorRef,
-    private elementRef: ElementRef,
-    private renderer: Renderer2
+    public changeDetector: ChangeDetectorRef,
+    public elementRef: ElementRef
   ) {
   }
 
   ngOnInit() {
-    this.workflowRunner = new WorkflowRunner(this);
+    this.workflow = new Workflow(this);
   }
 
   ngOnDestroy() {
-    this.workflowRunner.dispose();
+    this.workflow.dispose();
   }
 }
