@@ -1,6 +1,7 @@
 import { makeTest } from './scaffolding/runner';
+import { Misc } from './miscellaneous/misc';
 
-const checkViewport = (misc, viewportHeight) => {
+const checkViewport = (misc: Misc, viewportHeight: number) => {
   const position = misc.getScrollPosition();
   const size = misc.getScrollableSize();
   const bwdPadding = misc.padding.backward.getSize();
@@ -9,7 +10,7 @@ const checkViewport = (misc, viewportHeight) => {
   expect(position).toBeGreaterThan(bwdPadding);
   expect(position).toBeLessThan(size - fwdPadding);
   expect(bwdPadding + fwdPadding).toBeLessThan(size - viewportHeight);
-}
+};
 
 describe('Bug Spec', () => {
 
@@ -17,37 +18,38 @@ describe('Bug Spec', () => {
     const config = {
       datasourceName: 'default-delay-25',
       datasourceSettings: { startIndex: 1, bufferSize: 5, padding: 0.5 },
-      templateSettings: { viewportHeight: 100 }
+      datasourceDevSettings: { throttle: 50 },
+      templateSettings: { viewportHeight: 100 },
+      timeout: 4000
     };
 
     makeTest({
       title: 'should continue fetch',
       config,
-      it: (misc) => (done) => {
-
+      it: (misc: Misc) => (done: Function) => {
         const fwdCount = 4;
-        let wfCount = null, fetchCount;
+        let wfCount = null;
+        let jump = false;
 
-        spyOn(misc.workflowRunner, 'finalize').and.callFake(() => {
-          if (misc.workflowRunner.count < fwdCount) {
-            misc.scrollMax();
-          } else if (misc.workflowRunner.count === fwdCount) {
-            wfCount = misc.workflow.count;
-            misc.scrollMin();
+        spyOn(misc.workflow, 'finalize').and.callFake(() => {
+          if (misc.workflow.cyclesDone < fwdCount) {
+            setTimeout(() => misc.scrollMax(), 50);
+          } else if (misc.workflow.cyclesDone === fwdCount) {
+            wfCount = misc.workflow.cyclesDone;
+            setTimeout(() => misc.scrollMin(), 50);
             // 150 immediate jump will be here
-          } else if (misc.workflowRunner.count > fwdCount) {
+          } else if (misc.workflow.cyclesDone > fwdCount) {
             checkViewport(misc, config.templateSettings.viewportHeight);
             done();
           }
         });
 
-        spyOn(misc.workflow, 'finalize').and.callFake(() => {
-          if (misc.workflow.count === wfCount + 1) {
+        spyOn(misc.scroller, 'finalize').and.callFake(() => {
+          if (!jump && misc.workflow.cyclesDone === fwdCount) {
             misc.scrollTo(150);
+            jump = true;
           }
         });
-
-        misc.scrollMax();
       }
     });
   });
