@@ -1,150 +1,109 @@
 import { Direction } from '../src/component/interfaces';
 import { makeTest } from './scaffolding/runner';
+import { Misc } from './miscellaneous/misc';
 
-const min = 1, max = 100, scrollCount = 12;
+const min = 1, max = 100, scrollCount = 10;
 
 describe('EOF/BOF Spec', () => {
 
   const config = {
     bof: {
       datasourceName: 'limited',
-      datasourceSettings: { startIndex: min, bufferSize: 10, padding: 0.5 },
+      datasourceSettings: { startIndex: min, bufferSize: 10, padding: 0.5, adapter: true },
       templateSettings: { viewportHeight: 200 }
     },
     eof: {
       datasourceName: 'limited',
-      datasourceSettings: { startIndex: max - 10 + 1, bufferSize: 10, padding: 0.5 },
+      datasourceSettings: { startIndex: max - 10 + 1, bufferSize: 10, padding: 0.5, adapter: true },
       templateSettings: { viewportHeight: 200 }
     }
   };
 
-  const expectLimit = (misc, direction: Direction) => {
+  const expectLimit = (misc: Misc, direction: Direction, noscroll = false) => {
     const _forward = direction === Direction.forward;
     const elements = misc.getElements();
     expect(elements.length).toBeGreaterThan(config[_forward ? 'eof' : 'bof'].datasourceSettings.bufferSize);
     expect(misc.padding[_forward ? Direction.forward : Direction.backward].getSize()).toEqual(0);
-    expect(misc.padding[_forward ? Direction.backward : Direction.forward].getSize()).toBeGreaterThan(0);
+    if (!noscroll) {
+      expect(misc.padding[_forward ? Direction.backward : Direction.forward].getSize()).toBeGreaterThan(0);
+    }
     expect(misc.checkElementId(elements[_forward ? elements.length - 1 : 0], _forward ? max : min)).toEqual(true);
-    expect(misc.workflow.buffer.bof).toEqual(!_forward);
-    expect(misc.workflow.buffer.eof).toEqual(_forward);
+    expect(misc.scroller.buffer.bof).toEqual(!_forward);
+    expect(misc.scroller.buffer.eof).toEqual(_forward);
   };
 
-  describe('Begin of file', () => {
-    makeTest({
-      config: config.bof,
-      title: 'should get bof on init',
-      it: (misc) => (done) => {
-        expectLimit(misc, Direction.backward);
-        done();
-      }
-    });
+  const runLimitSuite = (eof = 'bof') => {
 
-    makeTest({
-      config: config.bof,
-      title: 'should reset bof after fwd scroll',
-      it: (misc) => (done) => {
-        expect(misc.workflow.buffer.bof).toEqual(true);
-        spyOn(misc.workflowRunner, 'finalize').and.callFake(() => {
-          expect(misc.workflow.buffer.bof).toEqual(false);
-          expect(misc.workflow.buffer.eof).toEqual(false);
-          done();
-        });
-        misc.scrollMax();
-      }
-    });
+    const isEOF = eof === 'eof';
 
-    makeTest({
-      config: config.bof,
-      title: 'should stop when bof is reached again',
-      it: (misc) => (done) => {
-        const count = misc.workflowRunner.count;
-        spyOn(misc.workflowRunner, 'finalize').and.callFake(() => {
-          if (misc.workflowRunner.count === count + 1) {
-            misc.scrollMin();
-          } else {
-            expectLimit(misc, Direction.backward);
+    describe((isEOF ? 'End' : 'Begin') + ' of file', () => {
+
+      const _eof = isEOF ? 'bof' : 'eof';
+      const direction = isEOF ? Direction.forward : Direction.backward;
+      const directionOpposite = isEOF ? Direction.backward : Direction.forward;
+      const doScroll = (misc: Misc) => isEOF ? misc.scrollMin() : misc.scrollMax();
+      const doScrollOpposite = (misc: Misc) => isEOF ? misc.scrollMax() : misc.scrollMin();
+
+      makeTest({
+        config: (<any>config)[eof],
+        title: `should get ${eof} on init`,
+        it: (misc: Misc) => (done: Function) =>
+          spyOn(misc.workflow, 'finalize').and.callFake(() => {
+            expectLimit(misc, direction, true);
             done();
-          }
-        });
-        misc.scrollMax();
-      }
-    });
+          })
+      });
 
-    makeTest({
-      config: config.bof,
-      title: 'should reach eof after some scrolls',
-      it: (misc) => (done) => {
-        const count = misc.workflowRunner.count;
-        spyOn(misc.workflowRunner, 'finalize').and.callFake(() => {
-          if (misc.workflowRunner.count < count + scrollCount) {
-            misc.scrollMax();
-          } else {
-            expectLimit(misc, Direction.forward);
-            done();
-          }
-        });
-        misc.scrollMax();
-      }
-    });
-  });
+      makeTest({
+        config: (<any>config)[eof],
+        title: `should reset ${eof} after scroll`,
+        it: (misc: Misc) => (done: Function) =>
+          spyOn(misc.workflow, 'finalize').and.callFake(() => {
+            const buffer = (<any>misc.scroller.buffer);
+            if (misc.workflow.cyclesDone === 1) {
+              expect(buffer[eof]).toEqual(true);
+              doScroll(misc);
+            } else {
+              expect(buffer[eof]).toEqual(false);
+              expect(buffer[_eof]).toEqual(false);
+              done();
+            }
+          })
+      });
 
-  describe('End of file', () => {
-    makeTest({
-      config: config.eof,
-      title: 'should get eof on init',
-      it: (misc) => (done) => {
-        expectLimit(misc, Direction.forward);
-        done();
-      }
-    });
+      makeTest({
+        config: (<any>config)[eof],
+        title: `should stop when ${eof} is reached again`,
+        it: (misc: Misc) => (done: Function) =>
+          spyOn(misc.workflow, 'finalize').and.callFake(() => {
+            if (misc.workflow.cyclesDone === 1) {
+              doScroll(misc);
+            } else if (misc.workflow.cyclesDone === 2) {
+              doScrollOpposite(misc);
+            } else {
+              expectLimit(misc, direction);
+              done();
+            }
+          })
+      });
 
-    makeTest({
-      config: config.eof,
-      title: 'should reset bof after bwd scroll',
-      it: (misc) => (done) => {
-        expect(misc.workflow.buffer.eof).toEqual(true);
-        spyOn(misc.workflowRunner, 'finalize').and.callFake(() => {
-          expect(misc.workflow.buffer.bof).toEqual(false);
-          expect(misc.workflow.buffer.eof).toEqual(false);
-          done();
-        });
-        misc.scrollMin();
-      }
+      makeTest({
+        config: (<any>config)[eof],
+        title: `should reach ${_eof} after some scrolls`,
+        it: (misc: Misc) => (done: Function) =>
+          spyOn(misc.workflow, 'finalize').and.callFake(() => {
+            if (misc.workflow.cyclesDone < scrollCount) {
+              doScroll(misc);
+            } else {
+              expectLimit(misc, directionOpposite);
+              done();
+            }
+          })
+      });
     });
+  };
 
-    makeTest({
-      config: config.eof,
-      title: 'should stop when eof is reached again',
-      it: (misc) => (done) => {
-        const count = misc.workflowRunner.count;
-        spyOn(misc.workflowRunner, 'finalize').and.callFake(() => {
-          if (misc.workflowRunner.count === count + 1) {
-            misc.scrollMax();
-          } else {
-            expectLimit(misc, Direction.forward);
-            done();
-          }
-        });
-        misc.scrollMin();
-      }
-    });
-
-    makeTest({
-      config: config.eof,
-      title: 'should reach bof after some scrolls',
-      it: (misc) => (done) => {
-        const count = misc.workflowRunner.count;
-        spyOn(misc.workflowRunner, 'finalize').and.callFake(() => {
-          if (misc.workflowRunner.count < count + scrollCount) {
-            misc.scrollMin();
-          } else {
-            expectLimit(misc, Direction.backward);
-            done();
-          }
-        });
-        misc.scrollMin();
-      }
-    });
-  });
+  runLimitSuite('bof');
+  runLimitSuite('eof');
 
 });
