@@ -1,22 +1,36 @@
 import { Direction } from '../src/component/interfaces';
-import { makeTest } from './scaffolding/runner';
+import { makeTest, TestBedConfig } from './scaffolding/runner';
+import { Misc } from './miscellaneous/misc';
+import { ItemsCounter } from './miscellaneous/itemsCounter';
 
-const singleForwardMaxScrollConfigList = [{
-  datasourceSettings: { startIndex: 100, bufferSize: 4, padding: 0.22 },
-  templateSettings: { viewportHeight: 71 },
+const configList: TestBedConfig[] = [{
+  datasourceSettings: { startIndex: 100, bufferSize: 4, padding: 0.22, itemSize: 20 },
+  templateSettings: { viewportHeight: 71, itemHeight: 20 },
   custom: { direction: Direction.forward, count: 1 }
 }, {
-  datasourceSettings: { startIndex: 1, bufferSize: 5, padding: 0.2 },
+  datasourceSettings: { startIndex: 1, bufferSize: 5, padding: 0.2, itemSize: 20 },
   templateSettings: { viewportHeight: 100 },
   custom: { direction: Direction.forward, count: 1 }
 }, {
-  datasourceSettings: { startIndex: -15, bufferSize: 12, padding: 0.98 },
-  templateSettings: { viewportHeight: 66 },
+  datasourceSettings: { startIndex: -15, bufferSize: 12, padding: 0.98, itemSize: 20 },
+  templateSettings: { viewportHeight: 66, itemHeight: 20 },
+  custom: { direction: Direction.forward, count: 1 }
+}, {
+  datasourceSettings: { startIndex: 1, bufferSize: 5, padding: 1, horizontal: true, itemSize: 90 },
+  templateSettings: { viewportWidth: 450, itemWidth: 90, horizontal: true },
+  custom: { direction: Direction.forward, count: 1 }
+}, {
+  datasourceSettings: { startIndex: -74, bufferSize: 4, padding: 0.72, horizontal: true, itemSize: 75 },
+  templateSettings: { viewportWidth: 300, itemWidth: 75, horizontal: true },
   custom: { direction: Direction.forward, count: 1 }
 }];
 
+configList.forEach(config => config.datasourceSettings.adapter = true);
+
+const treatIndex = (index: number) => index <= 3 ? index : (3 * 2 - index);
+
 const singleBackwardMaxScrollConfigList =
-  singleForwardMaxScrollConfigList.map(config => ({
+  configList.map((config, index) => ({
     ...config,
     custom: {
       ...config.custom,
@@ -25,11 +39,11 @@ const singleBackwardMaxScrollConfigList =
   }));
 
 const massForwardScrollsConfigList =
-  singleForwardMaxScrollConfigList.map((config, index) => ({
+  configList.map((config, index) => ({
     ...config,
     custom: {
       direction: Direction.backward,
-      count: 3 + index // 3-5 bwd scroll events per config
+      count: 3 + treatIndex(index) // 3-6 bwd scroll events per config
     }
   }));
 
@@ -38,7 +52,7 @@ const massBackwardScrollsConfigList =
     ...config,
     custom: {
       direction: Direction.backward,
-      count: 3 + index // 3-5 fwd scroll events per config
+      count: 3 + treatIndex(index) // 3-6 fwd scroll events per config
     }
   }));
 
@@ -47,7 +61,7 @@ const massBouncingScrollsConfigList_fwd =
     ...config,
     custom: {
       direction: Direction.forward,
-      count: (3 + index) * 2, // 3-5 (fwd + bwd) scroll events per config
+      count: (3 + treatIndex(index)) * 2, // 3-6 (fwd + bwd) scroll events per config
       bouncing: true
     }
   }));
@@ -57,7 +71,7 @@ const massBouncingScrollsConfigList_bwd =
     ...config,
     custom: {
       direction: Direction.backward,
-      count: (3 + index) * 2, // 3-5 (fwd + bwd) scroll events per config
+      count: (3 + treatIndex(index)) * 2, // 3-6 (fwd + bwd) scroll events per config
       bouncing: true
     }
   }));
@@ -67,7 +81,7 @@ const massTwoDirectionalScrollsConfigList_fwd =
     ...config,
     custom: {
       direction: Direction.forward,
-      count: (3 + index) * 2, // 3-5 fwd + 3-5 bwd scroll events per config
+      count: (3 + treatIndex(index)) * 2, // 3-6 fwd + 3-6 bwd scroll events per config
       mass: true
     }
   }));
@@ -77,12 +91,12 @@ const massTwoDirectionalScrollsConfigList_bwd =
     ...config,
     custom: {
       direction: Direction.backward,
-      count: (3 + index) * 2, // 3-5 fwd + 3-5 bwd scroll events per config
+      count: (3 + treatIndex(index)) * 2, // 3-6 fwd + 3-6 bwd scroll events per config
       mass: true
     }
   }));
 
-const doScrollMax = (config, misc) => {
+const doScrollMax = (config: TestBedConfig, misc: Misc) => {
   if (config.custom.direction === Direction.forward) {
     misc.scrollMax();
   } else {
@@ -90,70 +104,125 @@ const doScrollMax = (config, misc) => {
   }
 };
 
-const invertDirection = (config) => {
+const invertDirection = (config: TestBedConfig) => {
   const _forward = config.custom.direction === Direction.forward;
   config.custom.direction = _forward ? Direction.backward : Direction.forward;
 };
 
-const calculateIt = (config, misc) => {
-  const bufferSize = config.datasourceSettings.bufferSize;
-  const padding = config.datasourceSettings.padding;
-  const viewportSize = config.templateSettings.viewportHeight;
-  const direction = config.custom.direction;
-
-  const _forward = direction === Direction.forward;
-  const _paddingSize = misc.workflow.viewport.padding[direction].size;
-  const _edgeItemIndex = misc.workflow.buffer.getEdgeVisibleItem(direction).$index;
-
-  const sizeToFill = _paddingSize + padding * viewportSize;
-  const itemsToFill = Math.ceil(sizeToFill / misc.itemHeight);
-  const fetchCount = Math.ceil(itemsToFill / bufferSize);
-  const fetchedItemsCount = fetchCount * bufferSize;
-  const itemsToClip = fetchedItemsCount - itemsToFill;
-  const sizeToClip = itemsToClip * misc.itemHeight;
-  const edgeItemIndex = _edgeItemIndex + (_forward ? 1 : -1) * (fetchedItemsCount - itemsToClip);
-
-  return {
-    sizeToClip,
-    edgeItemIndex
-  };
+const getInitialItemsCounter = (misc: Misc): ItemsCounter => {
+  misc.fixture.detectChanges();
+  const { startIndex } = misc.scroller.settings;
+  const edgeItem = misc.scroller.buffer.getEdgeVisibleItem(Direction.forward);
+  const oppositeEdgeItem = misc.scroller.buffer.getEdgeVisibleItem(Direction.backward);
+  const result = new ItemsCounter();
+  result.set(Direction.forward, {
+    count: (<any>edgeItem).$index - startIndex + 1,
+    index: (<any>edgeItem).$index,
+    padding: 0
+  });
+  result.set(Direction.backward, {
+    count: startIndex - (<any>oppositeEdgeItem).$index,
+    index: (<any>oppositeEdgeItem).$index,
+    padding: 0
+  });
+  return result;
 };
 
-const shouldScroll = (config) => (misc) => (done) => {
-  const scrollCount = config.custom.count;
-  const count = misc.workflowRunner.count;
-  let result = calculateIt(config, misc);
+const getFullHouseDiff = (
+  viewportSize: number, paddingDelta: number, itemSize: number, bufferSize: number
+): number => {
+  const sizeToFill = viewportSize + 2 * paddingDelta; // size to fill the viewport + padding deltas
+  const itemsToFillNotRounded = sizeToFill / itemSize;
+  const itemsToFillRounded = Math.ceil(sizeToFill / itemSize);
+  const itemsToFill = itemsToFillRounded + (itemsToFillNotRounded === itemsToFillRounded ? 0 : 1);
+  const bufferSizeDiff = bufferSize - itemsToFill;
+  return Math.max(0, bufferSizeDiff);
+};
 
-  spyOn(misc.workflowRunner, 'finalize').and.callFake(() => {
-    if (misc.workflowRunner.count < count + scrollCount) {
-      const _forward = config.custom.direction === Direction.forward;
-      if (config.custom.bouncing) {
+const getCurrentItemsCounter = (misc: Misc, direction: Direction, previous: ItemsCounter): ItemsCounter => {
+  misc.fixture.detectChanges();
+  const { bufferSize, padding } = misc.scroller.settings;
+  const viewportSize = misc.scroller.viewport.getSize();
+  const itemSize = misc.scroller.buffer.averageSize;
+  const fwd = direction === Direction.forward;
+  const opposite = fwd ? Direction.backward : Direction.forward;
+  const delta = viewportSize * padding;
+
+  // handle direction (fetch)
+  const fullHouseDiff = getFullHouseDiff(viewportSize, delta, itemSize, bufferSize);
+  const _singleFetchCount = Math.ceil(delta / itemSize);
+  const singleFetchCount = Math.max(bufferSize, _singleFetchCount);
+  const itemsToFetch = previous.direction && previous.direction !== direction ?
+    (_singleFetchCount + fullHouseDiff) : singleFetchCount;
+  const newDirIndex = previous.get(direction).index + // previous edge index
+    (fwd ? 1 : -1) * (previous.get(direction).padding / itemSize) + // how many items are needed to fill the padding
+    (fwd ? 1 : -1) * itemsToFetch; // new pack of items
+
+  // handle opposite (clip)
+  const oppPadding = previous.get(opposite).padding;
+  const previousTotalSize = previous.total * itemSize + previous.paddings;
+  const sizeToClip = previousTotalSize - oppPadding - viewportSize - delta;
+  const itemsToClip = Math.floor(sizeToClip / itemSize);
+  const newOppIndex = previous.get(opposite).index + (fwd ? 1 : -1) * itemsToClip;
+  const newOppPadding = itemsToClip * itemSize + oppPadding;
+
+  const result = new ItemsCounter(direction);
+  result.set(direction, {
+    index: newDirIndex,
+    padding: 0
+  });
+  result.set(opposite, {
+    index: newOppIndex,
+    padding: newOppPadding
+  });
+  return result;
+};
+
+const shouldScroll = (config: TestBedConfig) => (misc: Misc) => (done: Function) => {
+  const custom = config.custom;
+  const wfCount = custom.count + 1;
+  const wfCountMiddle = Math.ceil(wfCount / 2);
+  let itemsCounter: ItemsCounter;
+
+  spyOn(misc.workflow, 'finalize').and.callFake(() => {
+    const cycles = misc.workflow.cyclesDone;
+    if (cycles === 1) {
+      itemsCounter = getInitialItemsCounter(misc);
+    } else {
+      itemsCounter = getCurrentItemsCounter(misc, custom.direction, itemsCounter);
+    }
+    if (cycles < wfCount) {
+      if (custom.bouncing) {
         invertDirection(config);
-      } else if (config.custom.mass) {
-        if (misc.workflowRunner.count - count === (scrollCount / 2)) {
+      } else if (custom.mass) {
+        if (cycles === wfCountMiddle) {
           invertDirection(config);
         }
-      }
-      if (misc.workflowRunner.count === count + scrollCount - 1) {
-        result = calculateIt(config, misc);
       }
       doScrollMax(config, misc);
     } else {
       // expectations
-      const direction = config.custom.direction;
-      const edgeItem = misc.workflow.buffer.getEdgeVisibleItem(direction);
-      expect(misc.padding[direction].getSize()).toEqual(result.sizeToClip);
-      expect(edgeItem.$index).toEqual(result.edgeItemIndex);
+      const direction = custom.direction;
+      const opposite = direction === Direction.forward ? Direction.backward : Direction.forward;
+      const edgeItem = misc.scroller.buffer.getEdgeVisibleItem(direction);
+      const oppositeEdgeItem = misc.scroller.buffer.getEdgeVisibleItem(opposite);
+      const edgeItemIndex = itemsCounter.get(direction).index;
+      const edgeOppositeItemIndex = itemsCounter.get(opposite).index;
+      expect(edgeItem && edgeItem.$index).toEqual(edgeItemIndex);
+      expect(oppositeEdgeItem && oppositeEdgeItem.$index).toEqual(edgeOppositeItemIndex);
+      expect((<any>misc.padding)[direction].getSize()).toEqual(itemsCounter.get(direction).padding);
+      expect((<any>misc.padding)[opposite].getSize()).toEqual(itemsCounter.get(opposite).padding);
+      expect(misc.checkElementContentByIndex(edgeItemIndex)).toEqual(true);
+      expect(misc.checkElementContentByIndex(edgeOppositeItemIndex)).toEqual(true);
       done();
     }
   });
-  doScrollMax(config, misc);
 };
 
 describe('Basic Scroll Spec', () => {
 
   describe('Single max fwd scroll event', () =>
-    singleForwardMaxScrollConfigList.forEach(config =>
+    configList.forEach(config =>
       makeTest({
         config,
         title: 'should process 1 forward max scroll',
@@ -233,3 +302,4 @@ describe('Basic Scroll Spec', () => {
   );
 
 });
+
